[♻️refactor]: KanbanBoard 내부 Issue 폴더를 생성하여 관심사에 따라 컴포넌트 분리
