+++ src/components/KanbanBoard/Issue/IssuePreviewList.tsx
@@ -3,8 +3,8 @@ import { css } from '@emotion/react';
 import { theme, ThemeType } from '@/styles/theme';
 import { RequirementStateType } from '@/types/requirement';
 
-import EllipsisIcon from '../Shared/Icons/EllipsisIcon';
-import RequirementCheckToggle from './Requirement/RequirementCheckToggle';
+import EllipsisIcon from '../../Shared/Icons/EllipsisIcon';
+import RequirementCheckToggle from '../Requirement/RequirementCheckToggle';
 
 type IssuePreviewListProps = {
   requirements: RequirementStateType[];
