+++ src/components/KanbanBoard/Issue/IssueTitle.tsx
@@ -7,8 +7,8 @@ import deleteStateAtom from '@/atoms/deleteStateAtom';
 import { useIssue } from '@/hooks/useIssue';
 import { IssueStateType } from '@/types/issue';
 
-import DeleteHoverBtn from '../Shared/Button/DeleteHoverBtn';
-import DeleteConfirmToast from '../Shared/Toast/DeleteConfirmToast';
+import DeleteHoverBtn from '../../Shared/Button/DeleteHoverBtn';
+import DeleteConfirmToast from '../../Shared/Toast/DeleteConfirmToast';
 import IssueInput from './IssueInput';
 
 type IssueTitleProps = {
