+++ src/components/KanbanBoard/KanbanDraggable.tsx
@@ -2,7 +2,7 @@ import { Draggable } from 'react-beautiful-dnd';
 
 import { IssueStateType } from '@/types/issue';
 
-import IssueMemo from './IssueMemo';
+import IssueMemo from './Issue/IssueMemo';
 
 type KanbanDraggableProps = {
   issue: IssueStateType;
